updating athena handler
