+++ stream_alert_cli/helpers.py
@@ -35,16 +35,6 @@ from stream_alert_cli.logger import LOGGER_CLI
 from stream_alert.rule_processor.firehose import FirehoseClient
 
 
-SCHEMA_TYPE_LOOKUP = {
-    bool: 'boolean',
-    float: 'float',
-    int: 'integer',
-    str: 'string',
-    dict: dict(),
-    list: list()
-}
-
-
 class MockCBAPI(object):
     """Mock for CbResponseAPI"""
 
@@ -95,33 +85,6 @@ class MockCBAPI(object):
             return MockCBAPI.MockBinary(banned=False, enabled=False, md5=file_hash)
         elif model == BannedHash:
             return MockCBAPI.MockBannedHash()
-
-
-def record_to_schema(record, recursive=False):
-    """Take a record and return a schema that corresponds to it's keys/value types
-
-    This generates a log schema that is compatible with schemas in conf/logs.json
-
-    Args:
-        record (dict): The record to generate a schema for
-        recursive (bool): True if sub-dictionaries should be recursed
-
-    Returns:
-        dict: A new record that reflects the original keys with values that reflect
-            the types of the original values
-    """
-    if not isinstance(record, dict):
-        return
-
-    result = {}
-    for key, value in record.iteritems():
-        # only worry about recursion for dicts, not lists
-        if recursive and isinstance(value, dict):
-            result[key] = record_to_schema(value, recursive)
-        else:
-            result[key] = SCHEMA_TYPE_LOOKUP.get(type(value), 'string')
-
-    return result
 
 
 def run_command(runner_args, **kwargs):
